pyproject.toml
[tool.poetry]
name = "hostfact-python-proxy"
version = "0.1.0"
description = ""
authors = ["Teun Ouwehand <teun@nextpertise.nl>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"
hostfact-python-client = "^0.1"
pydantic = {extras = ["dotenv"], version = "^2.2.1"}
requests = "^2.26.0"
pydantic-settings = "^2.0.3"
python-multipart = "^0.0.6"
pytest = "^7.4"
pytest-asyncio = "^0.21"
pytest-vcr = "^1.0"
httpx = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
